requirements.txt
# Astro Runtime includes the following pre-installed providers packages:
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# ── GCP Providers & SDKs ──────────────────────────────────────────
apache-airflow-providers-google>=10.0.0
google-cloud-bigquery>=3.20.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.19.0
google-cloud-dlp>=3.15.0
google-cloud-monitoring>=2.19.0
google-cloud-bigquery-storage>=2.24.0

# ── Apache Iceberg ────────────────────────────────────────────────
pyiceberg[glue,s3fs,gcsfs]>=0.6.0

# ── Data Validation ───────────────────────────────────────────────
soda-core-bigquery>=3.3.0

# ── Utilities ─────────────────────────────────────────────────────
pandas>=2.0.0
pyarrow>=15.0.0
smart-open[gcs]>=7.0.0
